# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YAOpenEvidence API",
    "description": "YAOpenEvidence 医学文献证据问答 API。\n\n- 问答任务异步执行：`POST /v1/answers` 返回 202，进度走 `GET /v1/jobs/{id}/events`（SSE）\n- 鉴权：管理员创建账号，登录后使用 `Authorization: Bearer <access_token>`（包括 SSE）\n- 错误：RFC 9457 Problem Details（`application/problem+json`），按 `code` 分支\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/answers": {
      "post": {
        "tags": [
          "answers"
        ],
        "summary": "创建问答任务",
        "operationId": "create_answer_v1_answers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "问答任务列表",
        "operationId": "list_answers_v1_answers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "ready",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AnswerSummary_"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/answers/{answer_id}": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "问答任务详情",
        "operationId": "get_answer_v1_answers__answer_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "answers"
        ],
        "summary": "删除终态答案",
        "operationId": "delete_answer_v1_answers__answer_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/answers/{answer_id}/markdown": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "完整渲染稿（Markdown）",
        "operationId": "get_answer_markdown_v1_answers__answer_id__markdown_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/answers/{answer_id}/papers/{n}/markdown": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "本次阅读原文快照（含段落锚点的 Markdown）",
        "operationId": "get_answer_paper_markdown_v1_answers__answer_id__papers__n__markdown_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/answers/{answer_id}/papers/{n}": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "本次阅读的第 n 篇（笔记、引文、段落、全文）",
        "operationId": "get_answer_paper_v1_answers__answer_id__papers__n__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerPaperDetail"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "ready",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "filters_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters Label"
          },
          "n_papers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Papers"
          },
          "n_fulltext": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Fulltext"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "question_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question En"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries",
            "default": []
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options",
            "default": {}
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/AnswerPaper"
            },
            "type": "array",
            "title": "Papers",
            "default": []
          },
          "body_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Md"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "kb_hits": {
            "items": {
              "$ref": "#/components/schemas/KbHit"
            },
            "type": "array",
            "title": "Kb Hits",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "question",
          "created_at"
        ],
        "title": "Answer"
      },
      "AnswerCreate": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "papers": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Papers",
            "default": 8
          },
          "years": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Years"
          },
          "year_from": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year From"
          },
          "year_to": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year To"
          },
          "quartiles": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Quartiles",
            "default": []
          },
          "journals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Journals",
            "default": []
          },
          "keep_unranked": {
            "type": "boolean",
            "title": "Keep Unranked",
            "default": false
          },
          "use_paywall": {
            "type": "boolean",
            "title": "Use Paywall",
            "default": true
          },
          "use_kb": {
            "type": "boolean",
            "title": "Use Kb",
            "default": true
          },
          "kb_hits": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Kb Hits",
            "default": 0
          },
          "max_chars": {
            "type": "integer",
            "maximum": 60000.0,
            "minimum": 4000.0,
            "title": "Max Chars",
            "default": 28000
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AnswerCreate",
        "description": "创建问答任务。年份既可用「最近 N 年」也可用明确区间，但不能同时给。"
      },
      "AnswerPaper": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "year": {
            "type": "string",
            "title": "Year",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "issn": {
            "type": "string",
            "title": "Issn",
            "default": ""
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "default": ""
          },
          "quartile": {
            "type": "string",
            "title": "Quartile",
            "default": ""
          },
          "rank_label": {
            "type": "string",
            "title": "Rank Label",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": [
              "pmc",
              "pdf",
              "inst",
              "abstract"
            ],
            "title": "Source",
            "default": "abstract"
          },
          "relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance"
          },
          "n_paragraphs": {
            "type": "integer",
            "title": "N Paragraphs",
            "default": 0
          },
          "n_citations": {
            "type": "integer",
            "title": "N Citations",
            "default": 0
          },
          "n_citations_verified": {
            "type": "integer",
            "title": "N Citations Verified",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "n"
        ],
        "title": "AnswerPaper"
      },
      "AnswerPaperDetail": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "year": {
            "type": "string",
            "title": "Year",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "issn": {
            "type": "string",
            "title": "Issn",
            "default": ""
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "default": ""
          },
          "quartile": {
            "type": "string",
            "title": "Quartile",
            "default": ""
          },
          "rank_label": {
            "type": "string",
            "title": "Rank Label",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": [
              "pmc",
              "pdf",
              "inst",
              "abstract"
            ],
            "title": "Source",
            "default": "abstract"
          },
          "relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance"
          },
          "n_paragraphs": {
            "type": "integer",
            "title": "N Paragraphs",
            "default": 0
          },
          "n_citations": {
            "type": "integer",
            "title": "N Citations",
            "default": 0
          },
          "n_citations_verified": {
            "type": "integer",
            "title": "N Citations Verified",
            "default": 0
          },
          "notes_md": {
            "type": "string",
            "title": "Notes Md",
            "default": ""
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/VerifiedQuote"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Facts",
            "default": []
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/Paragraph"
            },
            "type": "array",
            "title": "Paragraphs",
            "default": []
          },
          "fulltext_md": {
            "type": "string",
            "title": "Fulltext Md",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "n"
        ],
        "title": "AnswerPaperDetail"
      },
...
      "Citation": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "pid": {
            "type": "integer",
            "title": "Pid"
          },
          "sec": {
            "type": "string",
            "title": "Sec",
            "default": ""
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quotes",
            "default": []
          },
          "from_marker": {
            "type": "boolean",
            "title": "From Marker",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "n",
          "pid",
          "text"
        ],
        "title": "Citation",
        "description": "答案正文引到的一个原文段落。"
      },
...
      "JobError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "JobError"
      },
...
      "Page_AnswerSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnswerSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AnswerSummary]"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details。`code` 是稳定枚举，`detail` 仅供人读。",
        "properties": {
          "type": {
            "examples": [
              "urn:yaoe:error:not_found"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "examples": [
              "answer 'abc' not found"
            ],
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "examples": [
              "/v1/answers/abc"
            ],
            "title": "Instance",
            "type": "string"
          },
          "code": {
            "examples": [
              "not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "登录返回的可撤销会话令牌",
        "scheme": "bearer"
      }
    }
  }
}